added data caching and fixed linting errors
