+++ app-spec.yml
@@ -25,6 +25,6 @@ services:
       - path: /
     instance_count: 1
     instance_size_slug: basic-xxs
-    name: TCO2_Dashboard
-    run_command: gunicorn --worker-tmp-dir /dev/shm src.apps.TCO2_Dashboard.app:server
+    name: tco2_dashboard
+    run_command: gunicorn --worker-tmp-dir /dev/shm src.apps.tc02_dashboard.app:server
     source_dir: /
