+++ requirements.txt
@@ -4,9 +4,5 @@ millify
 pytest
 subgrounds==0.0.7
 web3
-dash==2.2.0
-numpy==1.22.2
-pandas==1.4.1
-plotly==5.6.0
 pycountry==22.1.10
