+++ src/apps/TCO2_Dashboard/Figures_Carbon_Pool.py
@@ -6,8 +6,9 @@ def deposited_over_time(df):
     df = df.sort_values(by="Date", ascending=True)
     df["Quantity"] = df["Quantity"].cumsum()
     fig = px.area(df, x="Date", y="Quantity")
-    fig.update_layout(paper_bgcolor=colors['bg_color'], plot_bgcolor=colors['bg_color'], font_color='white', xaxis=dict(showgrid=False),
-                      yaxis=dict(showgrid=False), hovermode='x unified',hoverlabel=dict(font_color='white', font_size=20), font_size=20)
+    fig.update_layout(paper_bgcolor=colors['bg_color'], plot_bgcolor=colors['bg_color'], font_color='white',
+                      xaxis=dict(showgrid=False), yaxis=dict(showgrid=False), hovermode='x unified',
+                      hoverlabel=dict(font_color='white', font_size=20), font_size=20)
     return fig
 
 
@@ -15,6 +16,7 @@ def redeemed_over_time(df):
     df = df.sort_values(by="Date", ascending=True)
     df["Quantity"] = df["Quantity"].cumsum()
     fig = px.area(df, x="Date", y="Quantity")
-    fig.update_layout(paper_bgcolor=colors['bg_color'], plot_bgcolor=colors['bg_color'], font_color='white', xaxis=dict(showgrid=False),
-                      yaxis=dict(showgrid=False), hovermode='x unified',hoverlabel=dict(font_color='white', font_size=20), font_size=20)
+    fig.update_layout(paper_bgcolor=colors['bg_color'], plot_bgcolor=colors['bg_color'], font_color='white',
+                      xaxis=dict(showgrid=False), yaxis=dict(showgrid=False), hovermode='x unified',
+                      hoverlabel=dict(font_color='white', font_size=20), font_size=20)
     return fig
